Bug 1525719: create a non-production /ducks/ endpoint for the React spike
This patch adds a new /ducks/ url just like our existing /docs/ url
for wiki documents, except it has a duck in the header. The URL route
is not installed if we're in production so this should be safe. We
are going to use this new url for our experimentation with React.