Bug 1522998: fix the CKEditor save buttons in maximized mode
The CKEditor save buttons are enabled and disabled depending
on whether any of the content has changed. The JavaScript to do this
uses a jQuery find() function to locate a ".editor-container" element
within a form, and then dispatches dirty or clean events on that
container element.

This is fine in the normal case, but if the editor is maximized
then the DOM hierarchy changes and there is no .editor-container.
The find() call returns an empty array and the events are dispatched
on the zero elements in that array.

This patch adds a special case: when no container element is found,
we dispatch the required events directly on the form (which is where
the handlers for those events are installed.)