bug 1505084: Move dependencies to devDependencies
npm dependencies were used in a non-standard way to distinguish between
tools needed to process assets for testing and deployment, versus tools
needed for local development and linting.

The more standard method is to split between runtime dependencies and
development dependencies, and currently none of ours are required during
runtime.