Bug 1525719: rewrite the MDN page header in React
This PR is an experiment with using React for the MDN page header.
This is a follow-on to PR #5254 which set up React for the MDN logo
for pages with special URLs that are never served in production. This
PR takes it further, adding the rest of the (logged-out) page header.

As part of the PR, we also:

- create a kuma/javascript/src/header directory for React components

- add a gettext.js module to (later) be used for l10n

- add a new .eslintrc.js config file for React-specific linting

- experiment with new visual design for the header

- Move the logo test to be in the same directory as the logo component.

- Add jest-emotion dev dependency for better snapshots of styles.

- Add snapshot tests for all the header-related components.