bug 1525719: Use new i18n domain for React
During development, the React views need all the strings from the
'javascript' domain, as well as some string that are currently in the
'django' domain. As we move toward production, strings may move as well,
but for now there is a lot of duplication.

Create a new 'react' domain that includes all the 'javascript' domain
strings, as well as React-specific strings, most or all of which are
repeated in the 'django' domain. Combine existing translations into a
compendium to populate the 'react' domain. Load the strings JS using a
new react_i18n function.

This is paired with a change to the locale repo to .gitignore react.po
and compendium files, until we're ready for independent translations.