Bug 1525719: move SVGs into external files
This patch modifies our React components that use SVGs so that the
SVGs live in external files rather than being inlined directly
into the components that use them.

And it also installs a webpack loader to automatically load them
as if they were React components. And it configures Jest to just
mock them out so they don't interfere with testing and so that changes
to the svg don't cause changes to the snapshots.