Bug 970278: Add noindex, nofollow tag to deprecated MindTouch pages
We've got various pages with slugs that begin with User:, Talk:,
User_talk:, and Template_talk: in our wiki that we really don't
want to support any more. We're not quite ready to delete them
but we at least don't want them to be turn up in search results.
This PR gives these pages a <meta name="robots" content="noindex,nofollow">
tag to prevent crawlers from indexing or crawling them.

Note that (per bug 1443179) we, in general, do not want to use
nofollow on pages. But in this case, since we're almost ready to
delete these pages, I don't think we want to rely on any links
they contain.