Bug 1525719: Query user data on the client side with new whoami API (#5289)
This PR introduces a new CurrentUser.Provider React component
that uses fetch() to get user data from the whoami API and then
uses the React Context API to share that user data with any components
that need it. The patch also includes changes to the existing Login
component to display the avatar of logged in users.

The patch also modifies the Login component to display a dropdown menu for
logged in users to allow view profile, edit profile and sign out commands.
This also includes modifications to the Dropdown component to make
it work with non-textual menu labels and to attach the menu to the
right side of the label instead of the left side.

Finally, the patch also bumps the minor version of React and makes
some changes to ensure that we can use Flow on our Jest tests.