Bug 1525719: implement the translations menu, plus UX fixes
This patch implement the translations menu in the MDN page header.
See also https://github.com/mdn/sprints/issues/1207
In order to implement it, this patch also modifies the document API
to return the local and translated names of the languages displayed
in the menu.

In addition, it modifies the Dropdown component so that it always
renders its children so that they are visible to web crawlers and
screen readers even when the menu is closed and they are not
visually rendered on screen.

The patch also improves the appearance of menus and the header
titlebar and TOC bar at the top of the page. The MDN logo is now
a standalone SVG file (we previously used a sprite file that included
an unnecessary mask).