Bug 1525719: show the list of contributors at the bottom of each page
This PR modifies the document API to include the list of contributors
and the last modified date for each document. On the frontend, it
modifies the React Page component to display that information near
the bottom of each page.

The patch also modifies the DocumentProvider component to add a
localeFromURL property to the document data. This new property is
intented as a single source of truth for the locale that is part
of the currently-displayed URL. Various React components that had
previously been manually extracting the locale from window.location
are modified to use this new localeFromURL property. This relatively
minor change required changes to test files and snapshots, making this
patch appear larger than it actually is.