bug 1538757: Disallow dynamic script detection
jQuery 2.x uses $.ajax() and variants $.get and $.post to load resources
via XHR. If the expected dataType is not specified, the response is used
to determine the type of data, and if it is JavaScript, it is
executed. This is useful for dynamically loading scripts, and an XSS
vector if a user can specify the URL.

This backports a jQuery 3.x mitigation, which adds a pre-filter to ajax
calls that disables treating cross-domain requests as scripts, as
suggested by jQuery authors (jQuery issue 2432)

If the caller specifies dataType='script', then a script load is
allowed, even if cross-domain. Current examples of dynamic script
loading are the browser compatibility table JS and the Stripe JS for the
payments banner. These already specify that a script is expected.