Bug 1525719: call history.pushState() on initial page
This patch fixes a bug in client-side navigation on the beta site.
We weren't calling history.pushState() on the first page load, so
after navigating away, the back button would not correctly return
to that initial page.

As part of the fix, the localeAndSlug property of the state is
modified to include an initial slash character.