bug 1475871: Test the (buggy) signup form template
The signup page includes an "Email address" form field that is a choice
field (emails returned from GitHub) plus an "Other" field for a custom
entry. The rendering of this page is untested, and the email field is
broken. This probably broke during the Django 1.8 to 1.11 upgrade, since
we're using some undocumented form / widget APIs.

Test the current functionality before fixing it.