Bug 1525719: make the React Page component responsive.
This patch modifies the page.jsx component to use CSS Grid layout
and display itself differently on narrow screens than it does on
larger screens. The main difference is that on narrow screens the
sidebar is moved to the bottom of the page, and the TOC bar is
hidden. There are also some font and margin size changes for
narrow screens as well.

The patch also includes a change to the react_document.html template
to make the MDN newsletter banner display with the correct margins.