Bug 1525719: Make the React header responsive. (#5331)
This patch modifies the React header component (and it subcomponents)
so that the layout works at phone, tablet and desktop size. It converts
the Header component from flexbox layout to grid layout and uses media
queries to select different grid templates at different sizes.

Additionally, media queries are used to set a base font size, and other
elements of the header (such as the MDN logo) are sized in 'em' units.