bug 1357416: Normalize slugs to NFC
This adds a NFC normalization step to the RevisionForm and DocumentForm,
used when editing, translating, and moving pages, so that slugs are
automatically normalized before saving.

Some unicode text is displayed the same, but is composed of a different
sequence of unicode codepoints. For example, the Greek small case
epsilon with a tonos (έ) can be represented by a single codepoint, or by
an epsilon (ε) followed by a combining acute accent. This can mean that
two unicode strings appear the same but are different because of the
code points.

Unicode defines normalization forms [1] that define the canonical
representation of these strings. One form, Normalization Form D (NFD,
or "decomposed"), uses the canonical decomposition, where έ is
represented by two codepoints, and there is a defined order for multiple
combining codepoints. Normalization Form C (NFC, or "composed") uses NFD
followed by a canonical composition, where έ is represented by one
codepoint.

The W3C specifies NFC for content [2], and it appears browsers perform
NFC on URLs. It is possible to enter a non-NFC slug in editing,
translating, and moving pages, and because the browser performs NFC,
these pages become unreachable. This is most likely to occur in the
Bengali (bn-BD) domain.

[1] http://www.unicode.org/faq/normalization.html
[2] https://www.w3.org/International/questions/qa-html-css-normalization