bug 1340194: No random passwords for acct recovery
The account recovery code, used to allow legacy Persona users to link
existing MDN profiles to GitHub, often sets a password for the user.

When generating a recovery URL, a random password was set on the user.
This may be because Django's PasswordResetForm checks that the user has
a usable password, and maybe there was other code that required that. I
checked with Django 1.11 and current code, and it is OK to send a
recovery email to a user with a blank or unusable password, so I'm
removing that code and adding test code that the URL can be generated.
Django made a similar choice, and allows password reset when the
existing password is unusable in Django 2.1.

When logging a user in via a recovery URL, the view set a random
password on the user, so that it could be used to log them in with
authenticate, and then set it to an unusable password. In Django 1.10,
login gained an optional backend parameter, which could be used to login
a user without credentials [2]. The code to add a temporary random password
has been removed.

[1] https://docs.djangoproject.com/en/2.1/topics/auth/passwords/#django.contrib.auth.hashers.is_password_usable
[2] https://docs.djangoproject.com/en/1.11/releases/1.10/#django-contrib-auth