rel='noopener' everywhere it's not (#5350)
I know this PR is a beast. It's hard to test beyond CI but I would appreciate some eyes on it for sanity. 

There were very few instances that were scary but some were a little bit. 
Links like `<a href="https://bugzilla.mocilla.org/xxx" target="_blank">` are low risk because they're only dangerous if bugzilla.mozilla.org gets compromised but it's good to make a separation because if any of those URLs it links to are compromised, it compromises us. 

I also noticed instance of `target="_blank"` in the .po files so I could use some help understanding what changes there mean and how to deal with them.