Fix jest tests to work in docker
I've been using jest and prettier locally and not really testing it
in docker. Ryan worked on a front-end patch recently and tried to
run tests and formatting from docker and uncovered some compatibility
issues. This PR fixes those issues, though in one case it simply
disables a test in docker.