Bug 1525719: move user and waffle-related ga() calls to client side
We currently hardcode a bunch of ga() analytics calls into the
HTML rendered by our Jinja templates, even when the data in question
depends on the current user or the current waffle flags. This means
that the HTML we generate can't be cached for re-use.

This PR modifies the react_document.html template so that our
React-based beta pages are cacheable and do not reference the
current user or current waffle flags. Instead, user and waffle
related ga() calls are now made on the client side in user-provider.jsx
when that data is obtained from the /whoami API.

This PR also adds ga('send', 'pageload') calls for client-side navigation.