Bug 1525719: add waffle flags to the /api/v1/whoami endpoint (#5386)
* Bug 1525719: add waffle flags to the /api/v1/whoami endpoint

For server-side rendered and cached versions of our pages, we need
to remove any user-specific data from the page. This includes waffle
flags, which can vary by user. This PR enables us to query the waffle
flags after a page has loaded, at the same time that we query the
user's name and profile picture.

Note that this PR does not remove the existing template code that
outputs waffle flags directly into a script in the HTML file. I'll
remove that later, after I've modified the client-side code to use the
dynamically queried flags.

* Address review comments; don't try to be clever about caching

* Improve tests with fake waffle data