Bug 1525719: change React localization to work with SSR (#5420)
This patch modifies the compile-react-i18n Makefile target so
that it generates JSON files instead of JS files for the strings
that are needed in the React front-end. Then, it modifies the
render_react_app template tag in ssr.py to read the localization
data for the specified locale and pass that data (which is mostly
the strings catalog) to the SSR server and to the client.

On the client-side, this patch introduces a new l10n.js module
that exports getLocale(), gettext(), ngettext(), and interpolate()
functions for localization. This new module replaces the old
locale-provider.jsx and gettext.js modules. These new gettext()
and ngettext() implementations are derived from the django
implementations.

index.jsx and ssr.jsx initialize the l10n module by calling its
localize() function with the necessary locale and localization data.