Bug 1525719: modify localization code to not violate CSP
In a recent patch I introduced a new l10n.js module for localization
of our React code. It used the Function() constructor to build a
plural function from the plural expression string that Django supplied.
Unfortunately, this violates the CSP policy that we would like to start
using.

So this PR modifies the backend to embed a valid plural function into
the window._react_data object that it includes in each HTML page. And
it also modifies ssr.jsx to use the Function() constructor to create
the plural function (in Node, so it does not violate CSP) when needed.

Because this patch decouples the plural expression from the string
catalog, it also removes the generic localizationData object and
uses the stringCatalog object directly.