Make the language menu work for React-based pages
The main.js file is not included into our React-based pages, so
the language menu in the footer does not autosubmit like it is
supposed to. This PR adds the required event handler to make it
work again. For now, at least, this code is just tacked on at
the end of index.jsx so it can be transpiled and bundled with
webpack.