Bug 1555004: stop including the react.js strings file into beta pages
For reasons I do not understand we've all-of-a-sudden started getting
500 errors from the server because the react.js strings files aren't
properly in the django staticfiles manifest. Rather than figure out
the source cause, this PR just removes the react.js include from the
templates that use it, because with the new React localization system
we don't actually use it anymore, and it should have been removed
a couple of weeks ago.

This is basically a partial revert of commit
bf5fad61ac58f22f5f3b7fcb712cd2f8f0e2dab9 which is what set up
that react.js file in the first place. React localization is now
handled during SSR with the react.json file. So this PR reverts
react.js back to the original javascript.js which provides localization
for any strings used by non-React JavaScript modules.  (When we remove
those non-React JavaScript modules, or at least all of them that use
strings, we'll be able to remove this javascript.js strings file
as well.