Convert the beta page menus back to hover behavior instead of click (#5454)
This PR updates the design of our menus and converts them back to
the original hover behavior, with a smaller target area so that they
don't pop up as easily. And it restores the links that were previously
available when clicking directly on the menu labels.

The hover behavior works with simulated touch events in chrome dev tools
but not in firefox dev tools. I'd like to just deploy this as is and then
we can more easily test with a real touch device.

Because this new menu does not respond to click events, I've had to delete
some test cases that were testing clicks on the menu.