Remove mdn.utils.randomString() (#5457)
PR #5445 causes a JS error in the developer console because a JS file
was is missing its the utils.js dependency. The problem is the
randomString() utility. Rather than include the entire utils.js file
this PR just removes the randomString() function and replaces the two
places it was called with a more efficient alternative.