Add Edit and History buttons to the titlebar for logged-in users.
Clicking these buttons takes the user to the wiki site. Currently
all logged-in users will see the buttons, but this PR includes some
code that makes a distinction between contributors and logged in
users who are not contributors. Once we've made that distinction
on the backend and in the whoami API, the frontend will show the
buttons to contributors onlyi.