Convert language menu to display language name instead of icon
This PR deletes the language.svg icon and changes the LanguageMenu
component to display the name of the current language instead of
that icon.

The PR also fixes a bug in the Dropdown menu component that was
requiring the user to click directly on the text of a menu item.
With the new (and simplified) CSS, a click anywhere within the
highlighted rectangle will work to trigger the menu item