Bug 1556324: Don't add an extra history entry when we load a page.
Our beta pages were calling History.pushState() when first loaded
(via an actual document load) in order to associate state with the
page so that we could go back to it. But this meant that that page
appeared twice in the history. This PR fixes that bug by switching
to History.replaceState().

This PR also fixes a related (unfiled) bug where we were not
updating the document title after client-side navigation so
the Back and Forward menus were showing the wrong document titles.