Bug 1557013: don't use zero width spaces for hyphenation
Some of our documents have really long one-word titles, like:

  CanvasRenderingContext2D.globalCompositeOperation

We want MDN to be able to break long strings like these reasonably.
In the past, we find the lower-to-upper case transitions and
inserted HTML <wbr> tags at those points to tell the browser that
they could break the word there. That was good user-facing UX, but
in bug 1523662 we discovered that for some reason Google was
sometimes showing the literal string "<wbr/>" in its search
results for our site.

So to fix that, we switched from <wbr> to zero width spaces
at the internal word transitions. This resulted in the same word
wrapping behavior, and google has been happy with it, I think.
But it caused this bug: that copy and paste is broken, especially
when trying to paste the title of a page into a code editor.

So in this iteration, we instead just using the CSS 'hyphens:auto'
property to tell browsers that they can break the long word. They
don't always break the word where we would like them to at the
internal word boundaries, but do seem to at least get syllable
boundaries (in English anyway). And, the benefit of this approach
is that browsers insert a hypen, so that users can more easily
tell that the line has been broken.

So we're running less code and we're getting a result that is
arguable better by just relying on the web platform to to the
right thing (or at least a reasonable thing).

Note that this PR adds the hyphens:auto change for both the new
React-based frontend and the old non-React frontend