Bug 1525719: make signin and signout work on the beta domain (#5483)
* Bug 1525719: make signin and signout work on the beta domain

Our login and logout endpoints require the ?next= redirect URL to
have the same origin as the endpoint. We were trying to use endpoints
on the wiki domain so that we didn't have to support them on the beta
domain. But that meant that when a user logged in or out, they were
taken away from the beta domain.

This PR fixes that by adding the auth endpoints to the beta domain. It
makes corresponding frontend changes in the login.jsx component.

* Remove unused signup endpoint from beta site.