Bug 1559480: make BCD tables appear and behave correctly on beta site.
BCD table formatting and behavior depends on code in
wiki-compat-trigger.js running on page load, so the tables don't
work correctly with client-side navigation. This PR is to fix that.

The first part is to make the BCD stylesheet part of the main bundle
that is loaded into every React-based page, so the styles are always
there and don't have to be loaded when we find a page with a BCD table.

The second part of this PR is the new bcd.js file that defines
a new activateBCDTables() function. We call this function with
useEffect() from the Article component so that it runs every time
we display a new article that might have a new BCD table in it.

activateBCDTables() is essentially a port of the jQuery code in
static/js/wiki-compat-table.js