Bug 1525719: improve error handling for search result pages (#5501)
On the beta site, if an error occured when fetching search results
we used to let it propagate. And then the router would handle it by
punting and doing a regular page reload. But search result pages aren't
rendered on the server, so a full page reload will end up making
the same API fetch call that just failed, and we're likely to end
up with an infinite reload loop.

This PR solves that problem: the SearchRoute fetch() method now
returns either a set of search results or an error, and the
SearchResultsPage component can display the results or the error.

Hopefully users will never see the error display. But if something
goes wrong (like elastic search is down) we'll get an error message
that should be helpful to diagnose the problem.