Bug 1561081: Don't display "Related Topics" heading if there are none (#5515)
Some pages don't have any related topics, and we shouldn't display
the "Related Topics" heading in the sidebar for those pages.

Note that this patch just removes the unnecessary heading from the
sidebar, but does not remove the sidebar itself. There may still be
a few pages htat have not related topics and also have no
sections to link to in an "On this page" section. Those pages have
no need of a sidebar, but they are still going to get one.