Port various JS modules to work with React and to remove jQuery. (#5503)
* Port various JS modules to work with React and to remove jQuery.

This large PR creates React-compatible versions of a number of
jQuery-based modules, and removes jQuery entirely from our beta pages.

This patch resolves bugs 1559480, 1560161, 1560135 and 1561037.

1) Bug 1559480: make BCD tables appear and behave correctly on beta site.

BCD table formatting and behavior depends on code in
wiki-compat-trigger.js running on page load, so the tables don't work
correctly with client-side navigation. This PR fixes that.

It makes the BCD stylesheet part of the main bundle that is loaded
into every React-based page, so the styles are always there and don't
have to be loaded when we find a page with a BCD table.

It also defines a new bcd.js file that defines an activateBCDTables()
function. We call this function with useEffect() from the Article
component so that it runs every time we display a new article that
might have a new BCD table in it.

activateBCDTables() is essentially a port of the jQuery code in
static/js/wiki-compat-table.js

2) Bug 1560161: syntax highlighting

In addition to patching up the BCD tables each time we display
a new article, we also need to perform syntax highlighting on
code blocks for each new article, so this commit adds a prism.js
file with a highlightSyntax() function that we can call from the
useEffect() hook of the Article component.

In addition it fixes what seems to be a regression in our SCSS files
where line highlighting was no longer happening on the beta site
or on the wiki site.

3) Bug 1560135: Open buttons for live examples

This PR ports the code from wiki-samples.js to live-examples.js so
that we can use an effect hook to add "Open in CodePen" and "Open in
JSFiddle" buttons to any documents that have live examples in them.

4) Bug 1561037: Narrow layouts for interactive examples

This PR ports the interactive example mediaQuery/narrowLayout code in
kuma/static/js/interactive.js to
kuma/javascript/src/interactive-examples.js and uses effect hooks in
article.jsx to get the code to run at the right time. This is so that
on narrow screens the interactive examples use their narrow layout
instead of their default wide layout.

5) Removing jQuery

This commit removes the last jQuery dependencies for our React-based pages.
It modifies two static/js/*.js files to convert simple jQuery calls
to the now widely-implemented standard alternatives. With those two
files modified (on both the wiki and beta domains) we no longer need
to include jQuery into the javascript bundle on the beta domain.

6) Other cleanup

Now that BCD and Prism JS is handled differently for the beta site
we no longer need window.mdn.assets output in our HTML, so that has
been removed for beta sites.

And, finally now that so much of the existing JS code has been ported
to the React codebase, there are no more gettext() calls in the
remaining legacy code used on the beta site, and so we can stop
outputting the javascript.js localization file.

* - check for existance of window object before using it so we
  don't crash the SSR server!