Bug 1525719: implement a React component for call-to-action ban… (#5516)
Our current system for call-to-action banners relies on the use of
waffle flags in our Jinja templates, which means that the rendered
html cannot be cached on a CDN. This PR introduces a <Banners>
component that can be used in place of the existing banners
javascript bundle and that gets waffle data from the user API
instead of relying on it in the Jinja template. The PR also set
up the react_homepage.html template and the landing_page.jsx
component to display the developer needs survey banner when the
waffle flag is turned on.