Bug 1561975: fix menu localization for beta pages. (#5520)
* Bug 1561975: fix menu localization for beta pages.

The header component defines menus in a static data structure. Originally
I had gettext() calls directly in that data structure. They were visible
to Puente, but the gettext() ran before string catalogs were available
so translations didn't happen even though they were available in the
catalog.

Then we switched to calling gettext() when rendering. That worked for
translations, but meant that the actual string literals were no longer
visible to Puente and weren't being picked up. So when the react.pot
file went live, we lost those string.

So then we moved the gettext() calls back to the static data structure
and are back to the original problem of calling gettext() before strings
are available.

This PR finally gets it right (I hope). The data structure that defines
the menus is created inside a useMemo() hook, so that it runs during
rendering (after strings are available). But because we are using
useMemo() we don't have to re-localize the strings every time we render
a header.

* - remove an extraneous gettext() call

* Fix the 'report a content problem' menu item url

* fix snapshot test