Fix a JS exception that was breaking client-side navigation.
In a recent PR I landed code that displayed document tags on
each page. In order to do this, I had to extend the doc API to
include a tags property in the JSON blob. It worked fine locally
but I forgot about our S3 cache of JSON blobs used in staging
and in production. The pre-rendered JSON blobs do not have the
new tags property, but the code for displaying the tags assumes
that it will always see tags or at least an empty array.

This was causing an exception during rendering and the router
fell back on doing a hard reload. This PR fixes the problem
by adding backwards compatibility code in the DocumentRoute class:
after fetching a JSON blob, we check for the tags property and
add one (an empty array) if the property does not exist.

With that fix in place, the frontend code can continue to
assume that the API always returns a tags array and the exception
goes away. Because of the way we do caching we will need to remember
to always add code like this whenever we evolve the doc schema in
the future.