Bug 1562271: don't scroll to top on popstate events
The router.jsx component for our beta pages scrolls to the top
every time it routes to a new page. But it should only do this for
new pages and not do it when the user navigates with the Back
or Forward buttons in the browser. This PR fixes this.