Don't show the loading bar for fast client-side navigation. (#5528)
Currently in our single-page app, we make a point of showing the
loading bar animation on every page transition, and even prolong
the length of the animation so that the user has postitive feedback
that something has happened.

We've gotten feedback that the loading bar makes client-side navigation
feel slower than it actually is, however. So this PR changes the
progress bar animation in two ways.

- first, it no longer artificially prolongs the animation
- second, it changes the animation so that nothing happens
  for the first 100ms.

The upshot is that when the page transition takes less than 100ms
the user will not see the progress bar at all. For navigations that
are under 200ms the user will just see a flash of blue, and for
navigations that take longer than that, they'll probably see something
that they recognize as a progress bar.

This is a trivial patch. We should try it out and see if we like
the site better with or without it.