Bug 1564405: don't display long, unwrapped excerpts (#5562)
This PR changes our search results page so that it ignores
result excerpts that are too long and so that it forces line
breaks even when excerpts don't have spaces. (There seems to
have been some kind of ElasticSearch change that made it start
sending us long excerpts without spaces.)

I'm sure there is still work to be done on the search results
page, but this fixes the particularly ugly problem.