Add a settings flag to use S3 for attachments
Why not Constance? Because we can't change the attachment's model at
runtime. We need a restart to completely switch out its storage backend.

Using settings implicitly enforces that requirement.

Note that the model gets cached at load, so we also can't switch storage
backends in the middle of test runs. It's possible to code around that,
but not worth the effort since we're dropping support for locally-served
attachments in the near future.