Remove Six
This was a manual process, and some changes were a little invasive:

- get_unique in kuma/core/utils.py

I believe the status quo ante was broken under Python 3, as it tried to
join bytes with `'\n'`, a string. I'm confident in the revision, but I'm
also fairly certain that this function is dead code and should be
removed in a subsequent pull request.

- generate_key in kuma/authkeys/models.py

Instead of calling `b''.join(chr(...) for ... in ...)`, I'm calling
`bytes(... for ... in ...)`, which consumes the iterable and
returns a joined bytestring.

Also note that the reference to Six in requirements/default_and_test.txt
needs to stay, as the version of Bleach we're using depends on Six.