Misc python dependency updates (#6404)
* Run poetry update

Both billiard and django-jinja have new upstream releases with miniscule
changes. Django-jinja's 2.6.0 release is just for dropping Python 2
support and confirming Python 3.8 support. Billiard makes a few changes
to fix a file descriptor leak, to update its test suite to pytest 5, and
to fix an uncaught exception when running in PyPy 2.x.

- Updating billiard (3.6.1.0 -> 3.6.2.0)
  https://github.com/celery/billiard/compare/3.6.1...3.6.2.0

- Updating django-jinja (2.5.0 -> 2.6.0)
  https://github.com/niwinz/django-jinja/blob/70c782a171253884fe6e50986cd9909229974737/CHANGES.adoc

* Unpin and update django-appconf (1.0.2 -> 1.0.3)

Django-appconf is a dependency of django-statici18n; we do not use it
directly. The only changes from 1.0.2 to 1.0.3 relate to dropping
support for older versions of Django and Python; our version of Django
is explicitly still supported.

- Updating django-appconf (1.0.2 -> 1.0.3)
  https://github.com/django-compressor/django-appconf/blob/1.0.3/docs/changelog.rst

* Update django-statici18n (1.7.0 -> 1.9.0)

No meaningful changes beyond fixing warnings and adding support for more
recent Python and Django versions.

- Updating django-statici18n (1.7.0 -> 1.9.0)
  https://github.com/zyegfryed/django-statici18n/blob/v1.9.0/docs/changelog.rst

* Remove django-flat-theme

This has been built into Django since version 1.9, released in 2015
https://docs.djangoproject.com/en/1.11/releases/1.9/#new-styling-for-contrib-admin

* Update django-csp (3.5 -> 3.6)

Only changes to keep up with new versions of Python, Django, and the CSP
spec itself.

- Updating django-csp (3.5 -> 3.6)
https://github.com/mozilla/django-csp/blob/2dfa6663f87c174e02c92fc83b1b2c5e07ba36df/CHANGES

* Update django-mysql (3.2.0 -> 3.3.0)

Just updates for new Python / Django versions, and one fix to
GroupConcat, which we do not use.

- Updating django-mysql (3.2.0 -> 3.3.0)
  https://github.com/adamchainz/django-mysql/blob/v3.3.0/HISTORY.rst

* Update django-storages (1.7.2 -> 1.9)

There are a few S3 related deprecations, but nothing we use or set.

- Updating django-storages (1.7.2 -> 1.9)
  https://github.com/jschneier/django-storages/blob/1.9/CHANGELOG.rst